Add files via upload
This algorithm was trained on a carefully selected sample of 225 field-tested 
examples and was specially tailored for each set of data. The sample was divided into two subsets, 85% for training and 15% for validation, in accordance with the approved approach. This set of training conditions was used for the 𝑉𝑠 and SPT collection of data. The determination coefficient (R) value is one of the most reliable indicators of the effectiveness of an ANN model. For the ANN model of the LI function, the R-values for training and testing on SPT data were 0.886 and 0.92, respectively, while the R values for training and testing on V𝑠 data were 0.89.